Refactor all food categories response and add SuccessDTO
